GEE refactored and follow new structure + added to main.py
